Split prose phase-output field recovery so detekt TooManyFunctions stops failing main.
ProsePhaseOutputParse hit the object threshold of 11 functions and failed Validate agent configs on every main push.
